Исправлены замечания.Убраны лишние импорты, выделен базовый класс для тестов.
