+++ src/test/java/ru/yandex/practicum/BROWSER.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum;
+
+public enum BROWSER {
+    CHROME,
+    FIREFOX,
+    SAFARI,
+    EDGE
+}
